+++ src/Exception/IndexNotFoundException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Locastic\Loggastic\Exception;
-
-final class IndexNotFoundException extends \Exception
-{
-}